Introduces flavors
