+++ app/src/bookdiary/java/dev/zezula/books/data/model/legacy/LegacyGroupShelfBookEntity.kt
