+++ app/src/bookdiary/java/dev/zezula/books/data/model/legacy/LegacyShelfEntity.kt
@@ -4,8 +4,8 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 // Schema:
-//SqliteColumn(name=_id, affinity=INTEGER, isNullable=true, inPrimaryKey=true)
-//SqliteColumn(name=name, affinity=TEXT, isNullable=true, inPrimaryKey=false)
+// SqliteColumn(name=_id, affinity=INTEGER, isNullable=true, inPrimaryKey=true)
+// SqliteColumn(name=name, affinity=TEXT, isNullable=true, inPrimaryKey=false)
 
 @Entity(tableName = "groups")
 data class LegacyShelfEntity(
