+++ app/src/main/java/dev/zezula/books/ui/screen/about/AboutDialog.kt
@@ -1,6 +1,5 @@
 package dev.zezula.books.ui.screen.about
 
-import androidx.compose.foundation.gestures.scrollable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
