+++ app/src/standard/java/dev/zezula/books/di/FlavoredAppModule.kt
@@ -1,0 +1,5 @@
+package dev.zezula.books.di
+
+import org.koin.dsl.module
+
+val flavoredAppModule = module {}
